backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.26.0
pyjwt[crypto]
python-dotenv
aiofiles